+++ db_assessment/optimusprime.py
@@ -14,8 +14,6 @@
 
 
 # Basic python built-in libraries to enable read, write and manipulate files in the OS
-import os
-import glob
 import sys
 
 # Manages command line flags and arguments
@@ -33,6 +31,9 @@ import rules_engine
 # Importing Optimus Prime Version
 import version
 
+# Import remote functionality
+from remote import runRemote
+
 # Information for analytics and tool improvement
 __version__= version.__version__
 
@@ -44,33 +45,6 @@ logging.getLogger().setLevel(level=logging.INFO)
 def getVersion():
 
     return __version__
-
-
-def get_id_token():
-    import google.auth
-    credentials, _ = google.auth.default()
-    credentials.refresh(google.auth.transport.requests.Request())
-    return credentials.id_token
-
-
-def runRemote(args):
-    import requests
-    id_token = os.getenv("ID_TOKEN") if os.getenv("ID_TOKEN") else get_id_token()
-    headers = {"Authorization": f"Bearer {id_token}"}
-    config = {
-        "projectId": args.projectname,
-        "dataset":  args.dataset,
-        "collectionId": args.collectionid
-    }
-    csvFilesLocationPattern = str(args.fileslocation) + '/*' + str(args.collectionid).replace(' ', '') + '.log'
-
-    # Getting a list of files from OS based on the pattern provided
-    # This is the default directory to have all customer database results from oracle_db_assessment.sql
-    files = import_db_assessment.getAllFilesByPattern(csvFilesLocationPattern)
-    files = {file_name: file_name for file_name in files}
-    result = requests.post(f"{args.remoteurl}/api/loadAssesment", files=files, data=config, headers=headers)
-    result.raise_for_status()
-    print(result.text)
 
 
 def runMain(args):
